objects/switch: merge setup and control in TRX
